POST-RAW requests: Don't parse response
When using the pseudo-method "POST-RAW", in additional to not parsing
the request body, also don't automatically parse the response body.
Instead, just return the unmodified response as plain string.

This is needed for requests which return a plain-text response and not
JSON.

Adapt the POI client which uses POST-RAW to parse as JSON itself.