Stateful Multigraph: Implement full support
Expand the Multigraph options and the Stateful Multigraph request
payload with all parameters supported nowadays.

All requests to the Stateful Multigraph must send a serviceUrl URL
parameter just like the Statistics Service on the production version.

When creating a new Multigraph, request its UUID as plain-text and don't
try to parse it as JSON.

Provide a tiles URL to be used as tile layer source for maps.